dev: add skill features

+++ player.cc
@@ -10,50 +10,50 @@ using namespace std;
 
 class player {
 private:
-    int base_hp;
-    int base_ep; // or sp
-    int base_sp;
-    int base_atk;
+    double base_hp;
+    double base_ep; // or sp
+    double base_sp;
+    double base_atk;
 
 
     
 public:
-    int cur_hp;
-    int cur_ep;
-    int cur_sp;
-    int cur_atk;
+    double cur_hp;
+    double cur_ep;
+    double cur_sp;
+    double cur_atk;
     int encouraged;
     int shell;
 
-    int get_base_hp() {
+    double get_base_hp() {
         return this->base_hp;
     }
-    void set_base_hp(int hp) {
+    void set_base_hp(double hp) {
         this->base_hp = hp;
         return;
     }
 
-    int get_base_ep() {
+    double get_base_ep() {
         return this->base_ep;
     }
-    void set_base_ep(int ep) {
+    void set_base_ep(double ep) {
         this->base_ep = ep;
         return;
     }
 
-    int get_base_sp() {
+    double get_base_sp() {
         return this->base_sp;
     }
-    void set_base_sp(int sp) {
+    void set_base_sp(double sp) {
         this->base_sp = sp;
         return;
     }
 
-    int get_base_atk() {
+    double get_base_atk() {
         return this->base_atk;
     }
 
-    void set_base_atk(int atk) {
+    void set_base_atk(double atk) {
         this->base_atk = atk;
         return;
     }
